Changed it so only one circle can be locked on to the cursor at a time, also the offset is memorized, so it doesn't just snap to the center of the cursor.
For some reason the diameter is used for the distance when seeing which circles to highlight. Whoever did the math seemed to have their radius and diameter mixed up. I did not fix that (will do that later probably, when I have time)

I increased the framerate limit to 60. Because 10 is way too slow and choppy. Hopefully this won't affect webpage performance (it runs at about 210-220 fps on chrome here)

Also, I changed the distance checks (if (sqrt((y1 - y0)^2 + (x1-x0)^2) < radius) to if ((y1 - y0)^2 + (x1-x0)^2 < radius^2), simply for efficiency. One could store radius squared if they wanted, say, a million circles instead of 20 or so.